% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters.R
\name{as_gemtc}
\alias{as_gemtc}
\title{Convert to a gemtc network}
\usage{
as_gemtc(x)
}
\arguments{
\item{x}{An \code{nma_data} object returned by \code{\link[=get_dataset]{get_dataset()}}.}
}
\description{
Convert to a gemtc network
}
