% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters.R
\name{as_netmeta}
\alias{as_netmeta}
\title{Convert to a netmeta object (arm- OR contrast-level)}
\usage{
as_netmeta(x)
}
\arguments{
\item{x}{An \code{nma_data} object returned by \code{\link[=get_dataset]{get_dataset()}}.}
}
\description{
Convert to a netmeta object (arm- OR contrast-level)
}
