% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_dataset}
\alias{get_dataset}
\title{Get a dataset by name}
\usage{
get_dataset(name)
}
\arguments{
\item{name}{dataset id from list_datasets()$id}
}
\value{
nma_data object
}
\description{
Get a dataset by name
}
