% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstream.R
\name{get_upstream_raw}
\alias{get_upstream_raw}
\title{Fetch a raw dataset from its upstream package}
\usage{
get_upstream_raw(id)
}
\arguments{
\item{id}{character}
}
\value{
the object as provided by the upstream package (no normalization).
}
\description{
Use ids like 'gemtc__smoking' that were auto-registered.
}
