% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstream.R
\name{list_upstream}
\alias{list_upstream}
\title{List upstream datasets only (registered but not embedded).}
\usage{
list_upstream()
}
\value{
tibble with at least id, package, object columns
}
\description{
List upstream datasets only (registered but not embedded).
}
