% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstream.R
\name{registry_path}
\alias{registry_path}
\title{Return the path to the package registry (installed or dev).}
\usage{
registry_path()
}
\description{
Return the path to the package registry (installed or dev).
}
