% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstream_scanner.R
\name{scan_and_register_upstream}
\alias{scan_and_register_upstream}
\title{Scan packages for likely NMA datasets and append to registry}
\usage{
scan_and_register_upstream(pkgs, install_if_missing = TRUE, write_now = TRUE)
}
\arguments{
\item{pkgs}{character vector of package names}

\item{install_if_missing}{logical, install from CRAN if needed}

\item{write_now}{logical, write to inst/registry/registry.csv}
}
\description{
Scan packages for likely NMA datasets and append to registry
}
